+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - More fine-grained control over the preprocessing, allows for multiple datasets.
+- Angle and grade are now continuous variables.
+- Loss is no longer calculated for the angle and grade.
 
 ### Removed